pyproject.toml
[project]
name = "muicebot-plugin-memory"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10, <=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "async_lru>=2.0.5",
    "jinja2>=3.1.6",
    "httpx>=0.28.1",
    "nonebot2>=2.4.1",
    "nonebot_plugin_alconna>=0.54.2",
    "nonebot_plugin_apscheduler>=0.5.0",
    "nonebot_plugin_orm>=0.7.7",
    "nonebot_plugin_localstore>=0.7.4",
    "numpy>=2.2.6",
    "muicebot>=1.0.1",
    "openai>=1.85.0",
    "sqlalchemy>=2.0.41",
]
authors = [
    { name = "Moemu", email = "i@snowy.moe" },
]

[project.optional-dependencies]
roberta = [
    "torch>=2.7.1",
    "transformers>=4.52.4",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"